Implement Warning the user after successful added if the series was 100 before to reach the maximum 10 digits SeriesNumber in RR

+++ Repository/ReceivingReportRepo.cs
@@ -13,7 +13,7 @@ namespace Accounting_System.Repository
             _dbContext = dbContext;
         }
 
-        public async Task<int> GetLastSeriesNumber()
+        public async Task<long> GetLastSeriesNumber()
         {
             var lastInvoice = await _dbContext
                 .ReceivingReports
